Cleanup whitespace, and only try conversion to unicode with PMJSONEncoder
Everything in Django should be assumed first as unicode. When trying to
convert an object down to a string that has a unicode character in it,
it will fail. And since fields from Django may very well have unicode
characters in them, this will fail consistently.

There's also no need to even try to convert to a string in this case.
Converting to unicode will work for plain strings too.